requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.4
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
oauthlib==3.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytube==10.0.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rope==0.18.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
